requirements.txt
aiogram==3.18.0
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
asyncpg==0.30.0
python-multipart==0.0.20
pydantic-settings==2.7.1
jinja2==3.1.5
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

